firmware-qcom-dragonboard410c: add dependency on xz
The self extracting archive internally uses xz compression, so when we unpack
the firmware archive, we end up using xz. HOSTTOOLS has been introduced in OE
core recently to avoid silently using tools from $PATH on the host. This has
reavealed that the firmware recipe was missing an explicit dependency on
xz-native tool (until now it was using it from the host!).

Reported-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>